backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
yfinance==0.2.33
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0 